+++ modals/DeleteOrderModal.tsx
@@ -21,12 +21,18 @@ const DelOrderModal: React.FC<Props> = ({ name, id, fetchData }) => {
 
   const handleDeleteOrder = async () => {
     try {
+      const accessToken = localStorage.getItem('accessToken');
+      if (!accessToken) {
+        console.error('No se encontró el token de acceso');
+        return;
+      }
       const response = await fetch(
         `${process.env.API_URL}/orders/${id}`,
         {
           method: 'DELETE',
           headers: {
-            'Content-Type': 'application/json' 
+            'Content-Type': 'application/json',
+            'Authorization': `Bearer ${accessToken}`
           },
         }
       );
